refactor: rename dataUri to mocks/uri
